Food, Store 데이터베이스 설계변경으로 리팩토링했습니다!

+++ src/main/java/com/example/Delivery/Store/SpringDataJPAStoreRepository.java
@@ -4,5 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SpringDataJPAStoreRepository extends JpaRepository<Store, Integer> {
+public interface SpringDataJPAStoreRepository extends JpaRepository<Store, Long> {
+    // 추가적인 메소드가 필요하다면 여기에 선언할 수 있습니다.
 }
